+++ src/Breadcrumbs.php
@@ -12,9 +12,10 @@ use yii\base\InvalidConfigException;
 use yii\helpers\ArrayHelper;
 
 /**
- * Button renders a bootstrap button.
+ * Breadcrumbs represents a bootstrap 4 version of [[\yii\widgets\Breadcrumbs]]. It displays
+ * a list of links indicating the position of the current page in the whole site hierarchy.
  *
- * For example,
+ * To use Breadcrumbs, you need to configure its [[links]] property, which specifies the links to be displayed. For example,
  *
  * ```php
  * echo Breadcrumbs::widget([